requirements.txt
python-decouple==3.8
websockets==14.2
